Remove import functionality and tests
