+++ src/clowder/error/clowder_yaml_error.py
@@ -29,17 +29,14 @@ class ClowderYAMLError(Exception):
 @unique
 class ClowderYAMLYErrorType(IntEnum):
     UNKNOWN = 99
-    RECURSIVE_IMPORT = 100
     REMOTE_NAME = 101
     MISSING_ENTRY = 102
     UNKNOWN_ENTRY = 103
     MISSING_YAML = 104
-    MISSING_IMPORTED_YAML = 105
     EMPTY_YAML = 106
     OPEN_FILE = 107
     INVALID_PROTOCOL = 108
     INVALID_REF = 109
-    MISSING_IMPORT = 110
     DEPTH = 111
     TYPE = 112
     MISSING_REPO = 113
