+++ src/clowder/util/clowder_utils.py
@@ -42,20 +42,6 @@ def existing_branch_projects(projects: List[Project], branch: str, is_remote: bo
     return any([p.existing_branch(branch, is_remote=is_remote) for p in projects])
 
 
-def get_clowder_yaml_import_path(import_name: str) -> str:
-    """Return path to imported clowder.yaml file
-
-    :param str import_name: Name of imported clowder.yaml
-    :return: Path to imported clowder.yaml file
-    :rtype: str
-    """
-
-    if import_name == 'default':
-        return os.path.join(ROOT_DIR, '.clowder', 'clowder.yaml')
-
-    return os.path.join(ROOT_DIR, '.clowder', 'versions', import_name, 'clowder.yaml')
-
-
 def filter_groups(groups: List[Group], names: List[str]) -> List[Group]:
     """Filter groups based on given group names
