--- src/clowder/yaml/__init__.py
@@ -1,3 +1,0 @@
-"""clowder.yaml __init__"""
-
-__MAX_IMPORT_DEPTH__ = 10
