--- src/clowder/yaml/validation/sources.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-"""clowder.yaml sources validation
-
-.. codeauthor:: Joe Decapo <joe@polka.cat>
-
-"""
-
-from clowder.yaml.validation.util import (
-    validate_empty,
-    validate_not_empty,
-    validate_optional_protocol,
-    validate_required_string,
-    validate_type
-)
-
-
-def validate_yaml_sources(sources: dict, yaml_file: str) -> None:
-    """Validate sources in clowder loaded from yaml file
-
-    :param dict sources: Parsed YAML python object for sources
-    :param str yaml_file: Path to yaml file
-    """
-
-    validate_type(sources, 'sources', list, 'list', yaml_file)
-    validate_not_empty(sources, 'sources', yaml_file)
-
-    for source in sources:
-        validate_type(source, 'source', dict, 'dict', yaml_file)
-        validate_not_empty(source, 'source', yaml_file)
-
-        args = ['name', 'url']
-        for arg in args:
-            validate_required_string(source, 'source', arg, yaml_file)
-
-        validate_optional_protocol(source, yaml_file)
-
-        validate_empty(source, 'source', yaml_file)
