+++ test/scripts/README.md
@@ -3,7 +3,6 @@
 ## [test_example_cats.sh](test_example_cats.sh)
 
 - Test invalid yaml files
-- Test `clowder.yaml` file with an import
 - Test all clowder commands
 - Test submodules
 
@@ -14,6 +13,11 @@
 
 ## [test_example_swift.sh](test_example_swift.sh)
 
+- Test forks
+
+## [test_example_misc.sh](test_example_misc.sh)
+
+- Test sources
 - Test forks
 
 ## [test_utilities.sh](test_utilities.sh)
