+++ test/scripts/cats/yaml.sh
@@ -14,9 +14,6 @@ echo "TEST: Test clowder yaml command"
 test_clowder_yaml() {
     print_single_separator
     echo "TEST: Test clowder yaml command"
-    pushd .clowder/versions || exit 1
-    test_cases=( $(ls -d import-*) )
-    popd || exit 1
 
     print_double_separator
     $COMMAND link || exit 1
@@ -24,13 +21,16 @@ test_clowder_yaml() {
     print_single_separator
     $COMMAND yaml -r || exit 1
 
-    for test in "${test_cases[@]}"
-    do
-        print_double_separator
-        $COMMAND link -v $test || exit 1
-        $COMMAND yaml || exit 1
-        print_single_separator
-        $COMMAND yaml -r || exit 1
-    done
+    # pushd .clowder/versions || exit 1
+    # test_cases=( $(ls -d import-*) )
+    # popd || exit 1
+    # for test in "${test_cases[@]}"
+    # do
+    #     print_double_separator
+    #     $COMMAND link -v $test || exit 1
+    #     $COMMAND yaml || exit 1
+    #     print_single_separator
+    #     $COMMAND yaml -r || exit 1
+    # done
 }
 test_clowder_yaml
