+++ test/scripts/test_example_cats.sh
@@ -33,5 +33,4 @@ print_double_separator
 "$TEST_SCRIPT_DIR/cats/repo.sh" || exit 1
 "$TEST_SCRIPT_DIR/cats/reset.sh" || exit 1
 "$TEST_SCRIPT_DIR/cats/yaml.sh" || exit 1
-"$TEST_SCRIPT_DIR/cats/yaml_import.sh" || exit 1
 "$TEST_SCRIPT_DIR/cats/help.sh" || exit 1
